tests/requirements_test.txt
-r ../requirements.txt

pytest==7.2.1
pytest-django==4.5.2
pytest-celery==0.0.0
pytest-mock==3.10.0
